+++ react-app/src/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
 import App from './App';
-import { ModalProvider } from './context/Modal';
+import { ModalProvider } from './context/ModalAndAuth';
 
 ReactDOM.render(
   <React.StrictMode>
